extract http requests helper
