+++ WireAPI/Tests/WireAPITests/ConnectionsAPITests.swift
@@ -108,8 +108,10 @@ class ConnectionsAPITests: XCTestCase {
 
         // checks we made the 3 correct requests
         for (index, receivedRequest) in httpClient.receivedRequests.enumerated() {
-            try await RequestSnapshotHelper<ConnectionsAPIBuilder>().verifyRequest(request: receivedRequest,
-                                                                                   resourceName: "v0.\(index)")
+            try await HTTPRequestSnapshotHelper().verifyRequest(
+                request: receivedRequest,
+                resourceName: "v0.\(index)"
+            )
         }
     }
 }
